+++ src/components/Icon/MinusIcon.tsx
@@ -10,9 +10,9 @@ function MinusIcon() {
       <path
         d="M1.5 1C6.18629 1 8.81371 1 13.5 1"
         stroke="#363636"
-        stroke-width="1.5"
-        stroke-linecap="round"
-        stroke-linejoin="round"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
       />
     </svg>
   );
